Add files via upload
Amidst the chaos, you notice that exactly one claim doesn't overlap by even a single square inch of fabric with any other claim. If you can somehow draw attention to it, maybe the Elves will be able to make Santa's suit after all!

For example, in the claims 3a, only claim 3 is intact after all claims are made.

What is the ID of the only claim that doesn't overlap?